add creatDB and dropDB
